Réunification des pages tribu en une page
