--- app/screens/AdminGroups/AdminGroups.scss
@@ -1,7 +1,0 @@
-.admin-groups-card {
-  padding: 10px;
-  margin: 20px;
-  max-width: 800px;
-  margin: auto;
-  text-align: center;
-}
