+++ app/screens/App/NavigationMenu/NavigationMenu.js
@@ -54,13 +54,7 @@ export default class NavigationMenu extends Component {
             <ListItemText primary="Validation des matchs" />
           </ListItem>
           <ListItem button onClick={this.goTo('/groups')}>
-            <ListItemText primary="Rejoindre une tribu" />
-          </ListItem>
-          <ListItem button onClick={this.goTo('/creategroup')}>
-            <ListItemText primary="Créer une tribu" />
-          </ListItem>
-          <ListItem button onClick={this.goTo('/admingroups')}>
-            <ListItemText primary="Administrer mes tribus" />
+            <ListItemText primary="Gestions des tribus" />
           </ListItem>
           <ListItem button onClick={this.goTo('/validinscription')}>
             <ListItemText primary="Valider l'inscription d'un membre" />
