--- app/screens/CreateGroup/CreateGroup.scss
@@ -1,37 +1,0 @@
-.create-group-container {
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  .create-group-card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 90%;
-    max-width: 800px;
-    margin-bottom: 15px;
-    padding: 15px;
-
-    .create-group-content {
-      display: flex;
-      align-items: center;
-      justify-content: space-around;
-      flex-wrap: wrap;
-      width: 90%;
-      padding: 0px;
-
-      .create-group-field {
-        margin: 10px;
-        width: 200px;
-      }
-    }
-
-    .create-group-card-action {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      width: 90%;
-    }
-  }
-}
