--- app/screens/CreateGroup/Loadable.js
@@ -1,8 +1,0 @@
-/**
- * Asynchronously loads the component
- */
-import CircularProgressLoadable from 'components/CircularProgressLoadable'
-
-export default CircularProgressLoadable({
-  loader: () => import('./index'),
-})
