+++ app/screens/Groups/AdminGroups/AdminGroupRow/index.js
