+++ app/screens/Groups/AdminGroups/AdminGroups.scss
@@ -1,0 +1,6 @@
+.admin-groups-card {
+  padding: 20px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
