+++ app/screens/Groups/CreateGroup/CreateGroup.scss
@@ -1,16 +1,20 @@
-.join-group-card {
+.create-group-card {
   padding: 20px;
   display: flex;
   flex-direction: column;
   align-items: center;
 
-  .join-group-content {
+  .create-group-content {
     display: flex;
     align-items: center;
     justify-content: space-around;
     flex-wrap: wrap;
     width: 90%;
-    margin: 10px 0px 10px 0px;
     padding: 0px;
+
+    .create-group-field {
+      margin: 10px ;
+      width: 200px;
+    }
   }
 }
