+++ app/screens/Groups/CreateGroup/CurrencyFormat/CurrencyFormat.jsx
