+++ app/screens/Groups/CreateGroup/CurrencyFormat/index.js
