+++ app/screens/Groups/CreateGroup/GroupCreateStatus/GroupCreateStatus.scss
