+++ app/screens/Groups/Groups.js
@@ -1,14 +1,18 @@
 import React from 'react'
 
 import JoinGroup from './JoinGroup'
+import CreateGroup from './CreateGroup'
+import AdminGroups from './AdminGroups'
 import MyGroups from './MyGroups'
 
 import './groups.scss'
 
 const Groups = () => (
   <div className="groups-container">
-    <JoinGroup />
     <MyGroups />
+    <JoinGroup />
+    <CreateGroup />
+    <AdminGroups />
   </div>
 )
