+++ app/screens/Groups/JoinGroup/GroupApplyStatus/GroupApplyStatus.scss
@@ -1,3 +1,7 @@
+.group-apply-snackbar {
+  max-width: !important 1000px;
+}
+
 .group-join-link {
   color: #7f7ff6;
 }
